+++ package.json
@@ -45,6 +45,7 @@
     "copy-to-clipboard": "^3.2.0",
     "ethereum-blockies-base64": "^1.0.2",
     "ethereumjs-tx": "^1.3.7",
+    "ethereumjs-util": "^7.0.10",
     "file-saver": "^2.0.0",
     "final-form": "^4.18.4",
     "final-form-arrays": "^3.0.1",
