+++ src/lib/constants.ts
@@ -74,9 +74,33 @@ const PROFILE_STATES = {
 
 const PROGRESS_ANIMATION_DELAY_MS = 500; // .animated-width
 
+const DEFAULT_HD_PATH = "m/44'/60'/0'/0/0";
+const ETH_ZERO_ADDR = '0x0000000000000000000000000000000000000000';
+const ETH_ZERO_ADDR_SHORT = '0x0';
+
+const WALLET_TYPES = {
+  MNEMONIC: Symbol('MNEMONIC'),
+  TICKET: Symbol('TICKET'),
+  SHARDS: Symbol('SHARDS'),
+  LEDGER: Symbol('LEDGER'),
+  TREZOR: Symbol('TREZOR'),
+  PRIVATE_KEY: Symbol('PRIVATE_KEY'),
+  KEYSTORE: Symbol('KEYSTORE'),
+  METAMASK: Symbol('METAMASK'),
+};
+
+const NONCUSTODIAL_WALLETS = new Set([
+  WALLET_TYPES.METAMASK,
+  WALLET_TYPES.TREZOR,
+  WALLET_TYPES.LEDGER,
+]);
+
 export {
   CHECK_BLOCK_EVERY_MS,
   DEFAULT_GAS_PRICE_GWEI,
+  DEFAULT_HD_PATH,
+  ETH_ZERO_ADDR,
+  ETH_ZERO_ADDR_SHORT,
   MAX_GAS_PRICE_GWEI,
   GAS_LIMITS,
   GEN_STATES,
@@ -88,10 +112,12 @@ export {
   MAX_STAR,
   MIN_PLANET,
   MAX_PLANET,
+  NONCUSTODIAL_WALLETS,
   PLANET_ENTROPY_BITS,
   STAR_ENTROPY_BITS,
   GALAXY_ENTROPY_BITS,
   SEED_ENTROPY_BITS,
+  WALLET_TYPES,
   ZOD,
   PROGRESS_ANIMATION_DELAY_MS,
 };
