+++ src/lib/useIsEclipticOwner.js
@@ -1,7 +1,7 @@
 import { useWallet } from 'store/wallet';
 import { usePointCache } from 'store/pointCache';
 
-import { eqAddr } from 'lib/wallet';
+import { eqAddr } from 'lib/utils/crypto';
 import useLifecycle from './useLifecycle';
 
 export default function useIsEclipticOwner() {
