+++ src/types/bridge-wallet.type.ts
@@ -1,0 +1,6 @@
+import { BIP32Interface } from 'bitcoinjs-lib';
+
+interface BridgeWallet extends BIP32Interface {
+  address: string;
+  passphrase: string;
+}
