+++ src/views/CancelTransfer.js
@@ -11,8 +11,7 @@ import * as need from 'lib/need';
 import { useLocalRouter } from 'lib/LocalRouter';
 import useCurrentPointName from 'lib/useCurrentPointName';
 import useEthereumTransaction from 'lib/useEthereumTransaction';
-import { GAS_LIMITS } from 'lib/constants';
-import { ETH_ZERO_ADDR } from 'lib/wallet';
+import { ETH_ZERO_ADDR, GAS_LIMITS } from 'lib/constants';
 import useLifecycle from 'lib/useLifecycle';
 
 import ViewHeader from 'components/ViewHeader';
