+++ src/views/Login/PrivateKey.js
@@ -4,7 +4,9 @@ import { Grid } from 'indigo-react';
 
 import { useWallet } from 'store/wallet';
 
-import { EthereumWallet, WALLET_TYPES, stripHexPrefix } from 'lib/wallet';
+import { EthereumWallet } from 'lib/wallet';
+import { stripHexPrefix } from 'lib/utils/crypto';
+import { WALLET_TYPES } from 'lib/constants';
 import useLoginView from 'lib/useLoginView';
 
 import FormError from 'form/FormError';
