+++ src/views/UrbitOS/NetworkingKeys.js
@@ -18,7 +18,7 @@ import {
 } from 'lib/keys';
 import useEthereumTransaction from 'lib/useEthereumTransaction';
 import { GAS_LIMITS } from 'lib/constants';
-import { addHexPrefix } from 'lib/wallet';
+import { addHexPrefix } from 'lib/utils/crypto';
 import useKeyfileGenerator from 'lib/useKeyfileGenerator';
 
 import ViewHeader from 'components/ViewHeader';
